feat(backend): put the scheduler and executable discovery behind osplat
executions_service shelled out to crontab and launchctl and chose the
launchd directories by sys.platform. The implementation moves verbatim
into osplat/_posix_scheduler.py, selected with launchd on macOS and
without on Linux, and Windows answers every kind as unsupported without
spawning anything (Task Scheduler is not wired). The Paths contract
gains the discovery questions the onboarding and plugin code kept
answering with os.name: executable candidates and executability
(PATHEXT on Windows), the login-PATH probe (none on Windows), the
backend entry name on disk (.exe), whether POSIX modes are enforced,
whether symlinks are available, and the shell used for a command
string. The tests exercise the POSIX scheduler module directly; the
allowlist drops executions_service.py. osplat modules now import on
Windows: the POSIX terminal modules and resource are imported lazily.
