+++ backend/agent_team_backend/osplat/_linux.py
@@ -15,7 +15,6 @@ from __future__ import annotations
 
 import logging
 import os
-import resource
 from pathlib import Path
 
 from ._posix import process_tree, terminal_backend
@@ -168,6 +167,8 @@ class LinuxResourceProbe:
         return "pss"
 
     def peak_rss_bytes(self) -> int | None:
+        import resource
+
         # Linux reports ru_maxrss in kilobytes.
         return resource.getrusage(resource.RUSAGE_SELF).ru_maxrss * 1024
 
@@ -210,6 +211,33 @@ class LinuxLayout(LinuxPaths):
     def askpass_launcher(self, helper_py: Path, python_exe: str | None) -> Path:
         return _posix_paths.askpass_launcher(helper_py, python_exe)
 
+    def executable_candidates(self, name: str) -> list[str]:
+        return _posix_paths.executable_candidates(name)
+
+    def is_executable(self, path: Path) -> bool:
+        return _posix_paths.is_executable(path)
+
+    def login_path_probe(self) -> list[str] | None:
+        return _posix_paths.login_path_probe()
+
+    def backend_entry_on_disk(self, entry: str) -> str:
+        return _posix_paths.backend_entry_on_disk(entry)
+
+    def enforces_posix_modes(self) -> bool:
+        return _posix_paths.enforces_posix_modes()
+
+    def symlinks_available(self) -> bool:
+        return _posix_paths.symlinks_available()
+
+    def shell_command(self, command: str) -> list[str]:
+        return _posix_paths.shell_command(command)
+
 
 paths = LinuxLayout()
 secret_files = _posix_secrets.secret_files
+
+from . import _posix_scheduler  # noqa: E402
+
+# crontab only: launchd does not exist here, so that kind lists as
+# unsupported without ever spawning `launchctl`.
+scheduler = _posix_scheduler.PosixScheduler(launchd=False)
